recommend/requirements.txt
asgiref==3.8.1
black==24.3.0
click==8.1.7
Django==5.0.3
djangorestframework==3.15.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
